requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
chromadb>=0.4.22
openai>=1.12.0
python-dotenv>=1.0.0
sentence-transformers>=2.5.1
